Implemented and replaced existing operator
This is an update to the original bar function operator, with more consistent results and an additional feature of being able toggle between horizontal and vertical orientation while visualizing data.
